function_app/requirements.txt
# Core Azure Functions
azure-functions>=1.16.0

# Core dependencies needed from main project
python-dotenv==1.0.0
pydantic==2.11.2

# Azure AI and Semantic Kernel
semantic-kernel==1.27.2
azure-ai-inference==1.0.0b9
azure-ai-projects==0.1.0
azure-core==1.33.0
azure-identity==1.14.0
azure-keyvault-secrets==4.9.0

# Data processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2

# For async operations
httpx==0.24.1
aiohttp==3.8.5

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
tqdm==4.66.1 